backend/requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
sqlalchemy==2.0.23
gunicorn==21.2.0
groq==0.28.0
python-dotenv==1.0.0
torch
torchvision
flask-cors==4.0.0
pandas>=1.5.0
pillow
